Rename TestkitTests - temp
